+++ releasenotes.md
@@ -6,6 +6,7 @@
 * The values of CVARs that have been changed from their defaults are now highlighted when entering the `cvarlist` CCMD in the console.
 * Voodoo dolls now move at the correct speed on very slow scrolling floors.
 * A bug is fixed whereby the BFG9000’s secondary projectiles would be fired in the wrong direction in some instances.
+* The `english` CVAR now uses a value of `british` rather than `international`.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)